+++ src/index.ts
@@ -2,4 +2,3 @@
  * Copyright (c) 2021 Interop Alliance and Dmitri Zagidulin. All rights reserved.
  */
 export { httpClientHandler, securityLoader } from './documentLoader.js'
-